move validators into appropiate settings
