+++ src/backend/InvenTree/plugin/base/barcodes/helper.py
@@ -28,20 +28,6 @@ def cache(func):
     return wrapper
 
 
-def barcode_plugins() -> list:
-    """Return a list of plugin choices which can be used for barcode generation."""
-    try:
-        from plugin import registry
-
-        plugins = registry.with_mixin('barcode', active=True)
-    except Exception:
-        plugins = []
-
-    return [
-        (plug.slug, plug.human_name) for plug in plugins if plug.has_barcode_generation
-    ]
-
-
 def generate_barcode(model_instance: InvenTreeBarcodeMixin):
     """Generate a barcode for a given model instance."""
     from common.settings import get_global_setting
